data-scripts/requirements.txt
olefile==0.44
Pillow==4.0.0
python-resize-image==1.1.3
python-slugify==1.2.1
PyYAML==3.12
requests==2.12.4
Unidecode==0.4.20
